refactor to awn controller
